pyproject.toml
[tool.ruff]
preview = true
line-length = 120
fix = true
output-format = "grouped"
extend-exclude = ["utilities/manifests"]

[tool.ruff.lint]
external = ["E501"]

[tool.ruff.format]
exclude = [".git", ".venv", ".mypy_cache", ".tox", "__pycache__", "utilities/manifests"]

[tool.mypy]
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
show_error_codes = true
warn_unused_ignores = true
ignore_missing_imports = true

[tool.uv.sources]

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
    "ipython>=8.12.3",
    "pre-commit>=4.3.0",
]

[project]
requires-python = "==3.13.*"
name = "opendatahub-tests"
version = "0.1.0"
description = "Tests repository for Open Data Hub (ODH)"
authors = []
license = {text = "Apache License (2.0)"}
readme = "README.md"
keywords = ["Openshift", "RHOAI", "ODH"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
]

dependencies = [
    "ipython>=8.18.1",
    "model-registry>=0.2.13",
    "openshift-python-utilities>=5.0.71",
    "pytest-dependency>=0.6.0",
    "pytest-progress",
    "python-simple-logger",
    "pyyaml",
    "tenacity",
    "types-requests>=2.32.0.20241016",
    "requests",
    "pytest-asyncio",
    "syrupy",
    "protobuf",
    "grpcio-reflection",
    "portforward>=0.7.1",
    "pytest-testconfig>=0.2.0",
    "pytest-jira>=0.3.21",
    "pygithub>=2.5.0",
    "timeout-sampler>=1.0.6",
    "shortuuid>=1.0.13",
    "jira>=3.8.0",
    "openshift-python-wrapper>=11.0.99",
    "semver>=3.0.4",
    "sqlalchemy>=2.0.40",
    "pytest-order>=1.3.0",
    "marshmallow==3.26.2,<4", # this version is needed for pytest-jira
    "pytest-html>=4.1.1",
    "fire",
    "llama_stack_client>=0.5.0,<0.6",
    "pytest-xdist==3.8.0",
    "dictdiffer>=0.9.0",
    "pytest>=9.0.0",
    "huggingface-hub>=1.2.3",
]

[project.urls]
homepage = "https://github.com/opendatahub-io/opendatahub-tests"
